+++ hal/x86_fsp_qemu_loader.c
@@ -40,7 +40,7 @@ void hal_prepare_boot(void)
 }
 #endif
 
-int hal_flash_write(uint32_t address, const uint8_t *data, int len)
+int hal_flash_write(uintptr_t address, const uint8_t *data, int len)
 {
     return 0;
 }
@@ -53,7 +53,7 @@ void hal_flash_lock(void)
 {
 }
 
-int hal_flash_erase(uint32_t address, int len)
+int hal_flash_erase(uintptr_t address, int len)
 {
     return 0;
 }
