+++ hal/zynq.c
@@ -1168,7 +1168,7 @@ void hal_prepare_boot(void)
 }
 
 /* Flash functions must be relocated to RAM for execution */
-int RAMFUNCTION hal_flash_write(uint32_t address, const uint8_t *data, int len)
+int RAMFUNCTION hal_flash_write(uintptr_t address, const uint8_t *data, int len)
 {
     return 0;
 }
@@ -1182,7 +1182,7 @@ void RAMFUNCTION hal_flash_lock(void)
 }
 
 
-int RAMFUNCTION hal_flash_erase(uint32_t address, int len)
+int RAMFUNCTION hal_flash_erase(uintptr_t address, int len)
 {
     return 0;
 }
